refactor: Update theme structure and UI components
- Moves `Theme` from `com.example.compose` to `com.example.lifehub.theme`.
- Moves `MainActivity` from `com.example.lifehub` to `com.example.lifehub.activity`.
- Updates `LifeHubTheme` to not default to system dark theme.
- Moves `color` package from `com.example.lifehub.ui.theme` to `com.example.lifehub.theme.color`.
- Updates `backgroundLight` color.
- Removes custom spacing from `HomeScreen` and `EmptyState` in favor of `LifeHubTheme.spacing`.
- Removes unused `EmptyState`.
