+++ app/src/main/java/com/example/lifehub/activity/MainActivity.kt
@@ -4,7 +4,7 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
-import com.example.compose.LifeHubTheme
+import com.example.lifehub.theme.LifeHubTheme
 import com.example.lifehub.navigation.NavGraph
 
 class MainActivity : ComponentActivity() {
