+++ app/src/main/java/com/example/lifehub/theme/color/Color.kt
@@ -18,7 +18,7 @@ val errorLight = Color(0xFFBA1A1A)
 val onErrorLight = Color(0xFFFFFFFF)
 val errorContainerLight = Color(0xFFFFDAD6)
 val onErrorContainerLight = Color(0xFF93000A)
-val backgroundLight = Color(0xFFF7F9FF)
+val backgroundLight = Color(0xFFF3F4F6)
 val onBackgroundLight = Color(0xFF171C21)
 val surfaceLight = Color(0xFFFCF8F8)
 val onSurfaceLight = Color(0xFF1C1B1B)
