+++ app/src/main/java/com/example/lifehub/ui/HomeScreen.kt
@@ -11,12 +11,12 @@ import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.unit.dp
+import com.example.lifehub.theme.LifeHubTheme
 import com.example.lifehub.util.UiStateScreenContainer
 
 @Composable
 fun HomeScreen() {
-    Scaffold(modifier = Modifier.fillMaxSize(),) { innerPadding ->
+    Scaffold(modifier = Modifier.fillMaxSize()) { innerPadding ->
         HomeScreenContent(modifier = Modifier.padding(innerPadding))
     }
 }
@@ -38,15 +38,13 @@ private fun HomeScreenContent(modifier: Modifier) {
                 .fillMaxSize()
                 .verticalScroll(rememberScrollState())
                 .padding(
-                    horizontal = 16.dp
+                    horizontal = LifeHubTheme.spacing.inset.medium
                 )
         ) {
             repeat(4) {
-                Column(modifier = Modifier.fillMaxWidth(), verticalArrangement = Arrangement.spacedBy(8.dp)) {
+                Column(modifier = Modifier.fillMaxWidth(), verticalArrangement = Arrangement.spacedBy(
+                    LifeHubTheme.spacing.stack.small)) {
                     Text(text = "Title category")
-//                    EmptyState(iconRes = R.drawable.baseline_attractions_24) {
-//                        // todo onclick
-//                    }
                 }
             }
         }
